Hot runner nozzle cooling method and its influence on plastic parts
Hot runner nozzle is one of important components of hot runner system. It is very important to control temperature of hot runner nozzle in injection mold. If temperature at the gate is too high, melt will stick to valve needle when mold is opened. If temperature at the gate is too low, melt at the gate will be too cold, which will make valve needle difficult to close.